refactor(racingcar): 뷰 렌더링을 ui 영역에 위임하여 도메인 모델과 뷰모델의 결합도를 낮춤
