+++ src/main/java/racingcar/RacingGameApplication.java
@@ -1,7 +1,7 @@
 package racingcar;
 
 import racingcar.domain.RacingGame;
-import racingcar.domain.RacingGameResultsBuilder;
+import racingcar.domain.RacingGameResults;
 import racingcar.ui.InputView;
 import racingcar.ui.ResultView;
 
@@ -9,9 +9,8 @@ public class RacingGameApplication {
 
     public static void main(String[] args) {
         InputView inputView = new InputView();
-
         RacingGame racingGame = new RacingGame(inputView.getTime(), inputView.getNumberOfCars());
-        RacingGameResultsBuilder gameResults = racingGame.game();
+        RacingGameResults gameResults = racingGame.game();
 
         ResultView resultView = new ResultView();
         resultView.print(gameResults);
